+++ src/game/board.cpp
@@ -2,7 +2,7 @@
 #include <3dstris/game/board.hpp>
 #include <3dstris/gui.hpp>
 
-Board::Board(const u32 width, const u32 height) : width(width), height(height) {
+Board::Board(const u16 width, const u16 height) : width(width), height(height) {
 	grid.reserve(width * height);
 	reset();
 }
@@ -26,23 +26,23 @@ PieceType Board::get(const Pos pos) const {
 	return get(pos.x, pos.y);
 }
 
-void Board::draw(const Vector2 origin, const u32 tileSize,
+void Board::draw(const Vector2 origin, const u8 tileSize,
 				 const float outerThick, const float gridThick) const {
 	GUI::drawOutline(Pos{origin.x, origin.y},
 					 WH{float(width * tileSize), float(height * tileSize)},
 					 outerThick, BOARD, 0.1f);
 
-	for (u32 y = 1; y < height; ++y) {
+	for (u16 y = 1; y < height; ++y) {
 		straightLine(Pos{origin.x, origin.y + y * tileSize},
 					 WH{float(width * tileSize), 0}, gridThick, GRID, 0);
 	}
-	for (u32 x = 1; x < width; ++x) {
+	for (u16 x = 1; x < width; ++x) {
 		straightLine(Pos{origin.x + x * tileSize, origin.y},
 					 WH{0, float(height * tileSize)}, gridThick, GRID, 0);
 	}
 
-	for (u32 y = 0; y < height; ++y) {
-		for (u32 x = 0; x < width; ++x) {
+	for (u16 y = 0; y < height; ++y) {
+		for (u16 x = 0; x < width; ++x) {
 			const PieceType p = get(x, y);
 
 			if (p != NONE && p != INVALID) {
@@ -60,9 +60,9 @@ void Board::draw(const Vector2 origin, const u32 tileSize,
 }
 
 void Board::clearLines() {
-	for (u32 y = 0; y < height; ++y) {
+	for (u16 y = 0; y < height; ++y) {
 		bool line = true;
-		for (u32 x = 0; x < width; ++x) {
+		for (u16 x = 0; x < width; ++x) {
 			if (get(x, y) == NONE) {
 				line = false;
 				break;
@@ -71,9 +71,9 @@ void Board::clearLines() {
 
 		if (line) {
 			_linesCleared++;
-			for (u32 curY = y; curY >= 1; --curY) {
-				for (u32 x = 0; x < width; ++x) {
-					set(x, curY, curY == 1 ? NONE : get(x, curY - 1));
+			for (u16 curY = y; curY >= 1; --curY) {
+				for (u16 x = 0; x < width; ++x) {
+					set(x, curY, curY == 1 ? NONE : get(x, u16(curY - 1)));
 				}
 			}
 		}
