+++ src/game/piece.cpp
@@ -71,7 +71,7 @@ bool Piece::collides(const int offX, const int offY) const {
 	return false;
 }
 
-void Piece::draw(const Vector2 origin, const u32 tileSize) const {
+void Piece::draw(const Vector2 origin, const u8 tileSize) const {
 	int ghostY = 0;
 	while (!collides(0, ++ghostY)) {
 	}
@@ -99,7 +99,7 @@ void Piece::draw(const Vector2 origin, const u32 tileSize) const {
 	}
 }
 
-void Piece::draw(const Vector2 origin, const u32 tileSize,
+void Piece::draw(const Vector2 origin, const u8 tileSize,
 				 const PieceShape& shape, const PieceType type) {
 	for (u8 y = 0; y < shape.size(); ++y) {
 		for (u8 x = 0; x < shape.size(); ++x) {
